[EI-133] [web] refactor: Add SelectedMetadataTitle component
