+++ web/app/ui/components/numerical-guidance/indicators-chart.tsx
@@ -4,11 +4,10 @@ import { useSelectedIndicatorBoardMetadata } from '@/app/business/hooks/use-sele
 import MultiLineChart from '../view/molocule/multi-line-chart';
 import { useMemo } from 'react';
 import Pending from '../view/molocule/pending';
-import EditableText from '../view/atom/editable-text';
-import { useDebouncedCallback } from 'use-debounce';
+import SelectedMetadataTittle from './selected-metadata-title';
 
 export default function IndicatorsChart() {
-  const { selectedMetadata, updateMetadata } = useSelectedIndicatorBoardMetadata();
+  const { selectedMetadata } = useSelectedIndicatorBoardMetadata();
   const { indciatorsValueViewModel, isPending } = useIndicatorsValueViewModel();
 
   const formattedIndicatorsRows = useMemo(
@@ -18,22 +17,10 @@ export default function IndicatorsChart() {
 
   const category = indciatorsValueViewModel?.tickerList ? indciatorsValueViewModel.tickerList : [];
 
-  const handleMetadataNameChange = useDebouncedCallback((name: string) => {
-    updateMetadata({
-      name,
-    });
-  }, 500);
-
   return (
     <>
       <Pending isPending={isPending}>
-        <EditableText
-          inputKey={selectedMetadata?.id}
-          onChange={handleMetadataNameChange}
-          readonly={selectedMetadata ? false : true}
-          className="w-40 text-lg font-medium p-0"
-          text={selectedMetadata ? selectedMetadata.name : 'No metadata'}
-        />
+        <SelectedMetadataTittle />
         <MultiLineChart
           data={formattedIndicatorsRows || []}
           categories={category}
